Migrate visualizer tabs to separate file
